Move ESD's validity checks into ReadOnlyStyledDocument
