+++ richtextfx/src/main/java/org/fxmisc/richtext/model/ReadOnlyStyledDocument.java
@@ -21,6 +21,7 @@ import org.reactfx.util.BiIndex;
 import org.reactfx.util.Either;
 import org.reactfx.util.FingerTree;
 import org.reactfx.util.FingerTree.NonEmptyFingerTree;
+import org.reactfx.util.Lists;
 import org.reactfx.util.ToSemigroup;
 import org.reactfx.util.Tuple2;
 import org.reactfx.util.Tuple3;
@@ -349,12 +350,20 @@ public final class ReadOnlyStyledDocument<PS, SEG, S> implements StyledDocument<
      */
     public Tuple3<ReadOnlyStyledDocument<PS, SEG, S>, RichTextChange<PS, SEG, S>, MaterializedListModification<Paragraph<PS, SEG, S>>> replace(
             int from, int to, UnaryOperator<ReadOnlyStyledDocument<PS, SEG, S>> mapper) {
+        ensureValidRange(from, to);
         BiIndex start = tree.locate(NAVIGATE, from);
         BiIndex end = tree.locate(NAVIGATE, to);
         return replace(start, end, mapper);
     }
 
-    Tuple3<ReadOnlyStyledDocument<PS, SEG, S>, RichTextChange<PS, SEG, S>, MaterializedListModification<Paragraph<PS, SEG, S>>> replace(
+    public Tuple3<ReadOnlyStyledDocument<PS, SEG, S>, RichTextChange<PS, SEG, S>, MaterializedListModification<Paragraph<PS, SEG, S>>> replace(
+            int paragraphIndex, int fromCol, int toCol, UnaryOperator<ReadOnlyStyledDocument<PS, SEG, S>> f) {
+        ensureValidParagraphRange(paragraphIndex, fromCol, toCol);
+        return replace(new BiIndex(paragraphIndex, fromCol), new BiIndex(paragraphIndex, toCol), f);
+    }
+
+    // Note: there must be a "ensureValid_()" call preceding the call of this method
+    private Tuple3<ReadOnlyStyledDocument<PS, SEG, S>, RichTextChange<PS, SEG, S>, MaterializedListModification<Paragraph<PS, SEG, S>>> replace(
             BiIndex start, BiIndex end, UnaryOperator<ReadOnlyStyledDocument<PS, SEG, S>> f) {
         int pos = tree.getSummaryBetween(0, start.major).map(s -> s.length() + 1).orElse(0) + start.minor;
 
@@ -390,6 +399,7 @@ public final class ReadOnlyStyledDocument<PS, SEG, S> implements StyledDocument<
      */
     public Tuple3<ReadOnlyStyledDocument<PS, SEG, S>, RichTextChange<PS, SEG, S>, MaterializedListModification<Paragraph<PS, SEG, S>>> replaceParagraph(
             int parIdx, UnaryOperator<Paragraph<PS, SEG, S>> mapper) {
+        ensureValidParagraphIndex(parIdx);
         return replace(
                 new BiIndex(parIdx, 0),
                 new BiIndex(parIdx, tree.getLeaf(parIdx).length()),
@@ -500,4 +510,24 @@ public final class ReadOnlyStyledDocument<PS, SEG, S> implements StyledDocument<
             }
         }
     }
+
+    private void ensureValidParagraphIndex(int parIdx) {
+        Lists.checkIndex(parIdx, getParagraphCount());
+    }
+
+    private void ensureValidRange(int start, int end) {
+        Lists.checkRange(start, end, length());
+    }
+
+    private void ensureValidParagraphRange(int par, int start, int end) {
+        ensureValidParagraphIndex(par);
+        Lists.checkRange(start, end, fullLength(par));
+    }
+
+    private int fullLength(int par) {
+        int n = getParagraphCount();
+        return getParagraph(par).length() + (par == n-1 ? 0 : 1);
+    }
+
+
 }
